% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeo2-package.R
\docType{package}
\name{hypergeo2-package}
\alias{hypergeo2}
\alias{hypergeo2-package}
\title{hypergeo2: Generalized Hypergeometric Function with Tunable High Precision}
\description{
Computation of generalized hypergeometric function with tunable high precision in a vectorized manner, with the floating-point datatypes from 'mpfr' or 'gmp' library. The computation is limited to real numbers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zhuxr11/hypergeo2}
  \item Report bugs at \url{https://github.com/zhuxr11/hypergeo2/issues}
}

}
\author{
\strong{Maintainer}: Xiurui Zhu \email{zxr6@163.com}

}
\keyword{internal}
